+++ manager/eris-mint/state/permissions_test.go
@@ -8,7 +8,7 @@ import (
 	"testing"
 	"time"
 
-	acm "github.com/eris-ltd/eris-db/manager/eris-mint/account"
+	acm "github.com/eris-ltd/eris-db/account"
 	"github.com/eris-ltd/eris-db/manager/eris-mint/evm"
 	ptypes "github.com/eris-ltd/eris-db/permission/types"
 	. "github.com/eris-ltd/eris-db/manager/eris-mint/state/types"
