+++ manager/eris-mint/state/state.go
@@ -6,7 +6,7 @@ import (
 	"io/ioutil"
 	"time"
 
-	acm "github.com/eris-ltd/eris-db/manager/eris-mint/account"
+	acm "github.com/eris-ltd/eris-db/account"
 	ptypes "github.com/eris-ltd/eris-db/permission/types"
 	. "github.com/eris-ltd/eris-db/manager/eris-mint/state/types"
 	txs "github.com/eris-ltd/eris-db/txs"
