+++ rpc/core/types/responses.go
@@ -1,7 +1,7 @@
 package core_types
 
 import (
-	acm "github.com/eris-ltd/eris-db/manager/eris-mint/account"
+	acm "github.com/eris-ltd/eris-db/account"
 	stypes "github.com/eris-ltd/eris-db/manager/eris-mint/state/types"
 	txtypes "github.com/eris-ltd/eris-db/txs"
 	"github.com/tendermint/tendermint/types"
