+++ test/testdata/testdata/testdata.go
@@ -1,7 +1,7 @@
 package testdata
 
 import (
-	"github.com/eris-ltd/eris-db/manager/eris-mint/account"
+	"github.com/eris-ltd/eris-db/account"
 	edb "github.com/eris-ltd/eris-db/erisdb"
 	ep "github.com/eris-ltd/eris-db/erisdb/pipe"
 	stypes "github.com/eris-ltd/eris-db/manager/eris-mint/state/types"
