+++ txs/tx_utils.go
@@ -2,7 +2,7 @@ package txs
 
 import (
 	"fmt"
-	acm "github.com/eris-ltd/eris-db/manager/eris-mint/account"
+	acm "github.com/eris-ltd/eris-db/account"
 	ptypes "github.com/eris-ltd/eris-db/permission/types"
 
 	"github.com/tendermint/go-crypto"
